requirements.txt
streamlit
streamlit-drawable-canvas
streamlit-autorefresh
Pillow
requests